refactor(commands): extract provider/agent commands into providers.rs
